Rename validate-question to question-validator
